terminadas migraciones, models y controllers, me logueo con mi primer usuario

+++ src/database/migrations/2025_12_05_164907_create_vacunos_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->integer('edad_estimada_meses')->nullable();
 
             // 4. OTROS DATOS
-            $table->float('peso_kg')->nullable();
+            $table->decimal('peso_kg', 8, 2)->nullable();
             $table->text('historial_notas')->nullable();
 
             $table->timestamps();
